// UsersCrud-Frontend/src/app/models/usuario.ts
export class usuario {
    id: number = 0;
    nombre: string = "";
    edad: number = 0;
    descripcion: string = "";
    apelido: string = "";
    email: string = "";
    fechaNacimiento: string = "";
    telefono: number = 0;
    idPaisResidencia: string = "";
    recibirInformacion: number = 1;
}